test_project/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
pycryptodome==3.11.0
requests==2.26.0